refactor some common logic from extract and data_analysis into data_reader
